Break down Results.tsx into CreateRecordSet and hooks
